merge and pull

+++ Models/PurchaseOrder.cs
@@ -9,7 +9,7 @@ namespace Accounting_System.Models
         [Display(Name = "PO No")]
         public string? PONo { get; set; }
 
-        public int SeriesNumber { get; set; }
+        public long SeriesNumber { get; set; }
         public DateTime Date { get; set; }
 
         [Display(Name = "Supplier Name")]
